Extract fixtures for android-instant-execution tests
