+++ subprojects/instant-execution/src/integTest/resources/org/gradle/instantexecution/builds/android-3.6-mini/build.gradle
@@ -2,12 +2,11 @@
 
 buildscript {
     repositories {
-        maven { url = "https://repo.gradle.org/gradle/ext-snapshots-local/" }
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.0-20190820155739+0200'
+        classpath 'com.android.tools.build:gradle:3.6.0-alpha6'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -16,7 +15,6 @@ buildscript {
 
 allprojects {
     repositories {
-        maven { url = "https://repo.gradle.org/gradle/ext-snapshots-local/" }
         google()
         jcenter()
     }
